moved trait to a new module
